Continue to move code to their new locations
